fix:修改分析逻辑

+++ js-backend/src/main/java/com/js/project/model/vo/InterfaceInfoVO.java
@@ -18,7 +18,7 @@ public class InterfaceInfoVO extends InterfaceInfo {
     /**
      * 调用次数
      */
-    private Integer totalNum;
+    private long totalNum;
 
     private static final long serialVersionUID = 1L;
 }
